+++ shrdr/src/core/qpbo.h
@@ -19,7 +19,7 @@
 namespace shrdr {
 
 using Time = uint32_t;
-using Dist = uint16_t;
+using Dist = uint32_t;
 
 template <
     class Cap,
